Tape memory buffers added. Modified the activation functions to use matrix dimensions instead of the full array size.
The way it is done now, there might be a potential memory issue because
the old nodes might not get cleared by the GC

This is something that I took care to avoid by making DM_rec's field
mutable at the lower level, but I am not sure the GC will do the right
thing here.

At any rate, it is no big deal.

The library is pretty fast right now. It might be worth optimizing the
final layers and adding batched gemm, but roughly, this is about it.

It might be possible to make it more memory efficient at a loss of speed
and greatly increased programming complexity.

Meh...if it comes to that, maybe somebody will write a Tensorflow
wrapper.